Refactor Find page UI into components.
