+++ frontend/src/features/find/ImageInput.js
@@ -3,80 +3,12 @@ import { jsx, css } from "@emotion/core"
 import tw, { styled } from "twin.macro"
 
 import React, { useState } from "react"
-import { useDispatch } from "react-redux"
-import { storeInputImage } from "./findSlice"
 
-const UploadImageComponent = props => {
-  const dispatch = useDispatch()
-
-  const handleImageUpload = e => {
-    dispatch(
-      storeInputImage({ uploadedImage: URL.createObjectURL(e.target.files[0]) })
-    )
-  }
-
-  return (
-    <div tw="text-gray-100 px-1 py-4">
-      <input
-        onChange={handleImageUpload}
-        tw="rounded text-gray-200"
-        type="file"
-      />
-    </div>
-  )
-}
-
-const CaptureImageComponent = props => {
-  return (
-    <div tw="text-gray-100 px-1 py-4">
-      <button
-        tw="rounded border border-white text-gray-100 hover:bg-gray-100 hover:font-semibold hover:text-gray-900 px-4 py-1 ml-2"
-        type="submit"
-      >
-        capture
-      </button>
-    </div>
-  )
-}
-
-const ImageLinkComponent = props => {
-  const { handleImageUpload } = props
-
-  return (
-    <div tw="text-gray-100 text-center px-1 py-4">
-      <form>
-        <input
-          onChange={handleImageUpload}
-          tw="rounded text-gray-900 p-1"
-          placeholder="enter link to image"
-        />
-        <button
-          tw="rounded border border-white text-gray-100 hover:bg-gray-100 hover:font-semibold hover:text-gray-900 px-4 py-1 ml-2"
-          type="submit"
-        >
-          search
-        </button>
-      </form>
-    </div>
-  )
-}
-
-const InputListButton = props => {
-  const { activeButton, ...others } = props
-
-  return (
-    <div>
-      {activeButton === true ? (
-        <button
-          tw="border-b-2 sm:border-l-2 border-gray-100 font-bold text-white hover:font-semibold p-1"
-          {...others}
-        />
-      ) : (
-        <button tw="text-gray-100 hover:font-semibold p-1" {...others} />
-      )}
-    </div>
-  )
-}
+// import components
+import UploadImageComponent from "./components/UploadImageComponent"
+import CaptureImageComponent from "./components/CaptureImageComponent"
+import ImageLinkComponent from "./components/ImageLinkComponent"
+import InputListButton from "./components/InputListButton"
 
 const ImageInput = () => {
   const [inputType, setInputType] = useState("upload_image")
